Added Time based unit tests

+++ src/net/ld/library/core/time/TimeSpan.java
@@ -33,13 +33,13 @@ public class TimeSpan {
 	// Variables
 	// =============================================
 
-	private long mTicks;
+	private double mTicks;
 
 	// =============================================
 	// Properties
 	// =============================================
 
-	public long ticks() {
+	public double ticks() {
 		return mTicks;
 	}
